:card_file_box: :pencil2: fix entity for products, fix price - change numeric for float

+++ src/products/entities/product.entity.ts
@@ -17,7 +17,7 @@ export class Product {
     title: string;
 
 
-    @Column('numeric', {
+    @Column('float', {
         default: 0
     })
     price: number
